requirements.txt
pillow>=9.0.0
httpx>=0.24.0 